화장실 리뷰 post API 구현

+++ src/main/kotlin/com/example/publictoilet_back/dto/ReviewResponseDto.kt
@@ -2,6 +2,6 @@ package com.example.publictoilet_back.dto
 
 import com.example.publictoilet_back.entity.Review
 
-class ReviewResponseDto(val score : Double, val comment : String?) {
+class ReviewResponseDto(val score : Float, val comment : String?) {
     constructor(entity : Review) : this(entity.score, entity.comment)
 }
